Cleanup on Forms
Switching to using JSON schema layout for ngx-formly. Makes it simpler
and there isn't much need for more complex form layout at this point.
May support advanced forms in the future that will allow this. But for
attaching form data to documents, it makes sense for the layout to be a
single column anyways.