Can define and sort form submission JSON fields in data tables
Implemented in data table component, so this will technically work
anywhere, but is geared towards dynamically adding columns to the
table that are JSON fields.

Also fix sort configuration cache so it uses the prop names used
by ngx-datatable. Was causing issues when the column name was
different than the prop ngx-datatable assigned it (they lower-camelize
the column names).