Progress on a big overhaul to data tables
Data tables are now mostly a wrapper for ngx-datatable template generation.
Added a new component that wraps a data table and binds filters, sorting,
and pagination to the activated route query parameters. This way we can do
modal selects without having to deal with the headache of a named outlet.
Just wrap the data table in a controlling component that provides the
data via a service object that implements an index() method.

Lots of work remaining, including probably defining interfaces and types
(using TypeScript, after all), as well as fixing the remaining broken usage(s)
of data tables.