Re-enable route-bound data table filters
Similar to sorts, column identifiers are used instead of the actual
expressions that build them, to keep web links from exploding in size from
generated column expressions. This mechanism is a pretty bland string
replacement, but can cause issues if the string to replace appears in a
string literal in the expression. Not something to expect to occur commonly,
but a concern nonetheless.